Application working jar file
I have included the jar file of the application. This will be able to be run on linux, windows and mac os. I have not included the exe file (includes the icon) because I would need to upload the runtime folder for it to work which is greater than 100 MB after compression and github doesn't allow it.